+++ extension/src/ALObject/ALParser.ts
@@ -213,7 +213,7 @@ function matchALControl(parent: ALControl, lineIndex: number, codeLine: ALCodeLi
 }
 
 function getProperty(parent: ALControl, lineIndex: number, codeLine: ALCodeLine) {
-    let propertyResult = codeLine.code.match(/^\s*(?<name>ObsoleteState|SourceTable|PageType|ApplicationArea)\s*=\s*(?<value>"[^"]*"|[\w]*);/i);
+    let propertyResult = codeLine.code.match(/^\s*(?<name>ObsoleteState|SourceTable|PageType|ApplicationArea|Access)\s*=\s*(?<value>"[^"]*"|[\w]*);/i);
     if (propertyResult && propertyResult.groups) {
         let property = new ALProperty(parent, lineIndex, propertyResult.groups.name, propertyResult.groups.value);
         return property;
